Add --sort-by flag to threads sub-command
