+++ rfd/VERSION
@@ -1,1 +1,1 @@
-0.3.6
+0.4.0
