+++ rfd/__init__.py
@@ -6,3 +6,5 @@ __title__ = "RFD CLI"
 __author__ = "Dave Gallant"
 __license__ = "Apache 2.0"
 __copyright__ = "(c) 2018 Dave Gallant"
+
+API_BASE_URL = "https://forums.redflagdeals.com"
