+++ rfd/posts.py
@@ -1,0 +1,8 @@
+# pylint: disable=old-style-class
+
+
+class Post:
+    def __init__(self, body, score, user):
+        self.body = body
+        self.score = score
+        self.user = user
