+++ rfd/utils.py
@@ -1,4 +1,18 @@
 """This module provides utility functions that are used within rfd"""
+try:
+    from urllib.parse import urlparse  # python 2
+except ImportError:
+    from urlparse import urlparse  # python 1
+from bs4 import BeautifulSoup
+
+
+def strip_html(text):
+    return BeautifulSoup(text, "html.parser").get_text()
+
+
+def is_valid_url(url):
+    result = urlparse(url)
+    return all([result.scheme, result.netloc, result.path])
 
 
 def is_int(number):
